Implemented 'add_to_spreadsheet' task
